@@ -6,7 +6,7 @@ description: Demonstrates how to use the vdotok_stream plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,14 +22,14 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  page_transition: ^1.1.7+3
+  page_transition: "^2.0.5"
   flutter_statusbarcolor: ^0.2.3
-  flutter_ringtone_player: ^2.0.0
-  vibration: ^1.7.3
+  flutter_ringtone_player: ^3.2.0
+  vibration: ^1.7.4-nullsafety.0
   equatable:
   shared_preferences: ^2.0.5
-  http: ^0.12.2
-  json_annotation: ^3.1.1
+  http: ^0.13.4
+  json_annotation: ^4.5.0
   device_info: ^2.0.0
   path_provider: ^2.0.1
   provider: ^5.0.0
@@ -45,6 +45,7 @@ dependencies:
   # flutter_ip: ^0.1.4
   wifi_info_flutter: ^2.0.2
   wakelock: ^0.5.6
+  web_socket_channel: ^2.2.0
   #proximity_sensor: ^1.0.2
   #screen: ^0.0.5
   audioplayers: 
