@@ -29,16 +29,16 @@ class AuthProvider with ChangeNotifier {
   Status get loggedInStatus => _loggedInStatus;
   Status get registeredInStatus => _registeredInStatus;
 
-  User _user = new User();
+  User _user = new User(auth_token: '', full_name: '');
   User get getUser => _user;
 
   SharedPref _sharedPref = SharedPref();
-  String _loginErrorMsg;
+  late String _loginErrorMsg;
   String get loginErrorMsg => _loginErrorMsg;
 
-  String _registerErrorMsg;
+  late String _registerErrorMsg;
   String get registerErrorMsg => _registerErrorMsg;
-  String _completeAddress;
+  late String _completeAddress;
   String get completeAddress => _completeAddress;
 
   Future<bool> register(String email, username, password) async {
@@ -78,7 +78,7 @@ class AuthProvider with ChangeNotifier {
       "device_model": model,
       "device_os_ver": version,
       "app_version": "1.1.5",
-      "project_id":project_id
+      "project_id": project_id
     };
     final response = await callAPI(jsonData, "SignUp", null);
     print("this is response $response");
@@ -88,7 +88,7 @@ class AuthProvider with ChangeNotifier {
       notifyListeners();
       return false;
     } else {
-      _completeAddress  = response['media_server_map']['complete_address'];
+      _completeAddress = response['media_server_map']['complete_address'];
       SharedPref sharedPref = SharedPref();
       sharedPref.save("authUser", response);
       _registeredInStatus = Status.Registered;
@@ -106,7 +106,7 @@ class AuthProvider with ChangeNotifier {
     Map<String, dynamic> jsonData = {
       "email": username,
       "password": password,
-      "project_id":project_id
+      "project_id": project_id
     };
     final response = await callAPI(jsonData, "Login", null);
     print("this is response $response");
@@ -115,7 +115,7 @@ class AuthProvider with ChangeNotifier {
       _loginErrorMsg = response['message'];
       notifyListeners();
     } else {
-     _completeAddress  = response['media_server_map']['complete_address'];
+      _completeAddress = response['media_server_map']['complete_address'];
       print("this is complete address ${_completeAddress}");
       SharedPref sharedPref = SharedPref();
       sharedPref.save("authUser", response);
@@ -129,7 +129,7 @@ class AuthProvider with ChangeNotifier {
     SharedPref sharedPref = SharedPref();
     sharedPref.remove("authUser");
     _loggedInStatus = Status.LoggedOut;
-    _user = null;
+    // _user = /;
     notifyListeners();
   }
 
