@@ -5,7 +5,7 @@ import '../../../src/core/services/server.dart';
 enum ContactStates { Loading, Success, CreateGroup }
 
 class ContactProvider with ChangeNotifier {
-  ContactList _contactList;
+  late ContactList _contactList;
   ContactList get contactList => _contactList;
   ContactStates _contactStates = ContactStates.Loading;
   ContactStates get contactState => _contactStates;
