@@ -1,13 +1,12 @@
 import 'package:flutter/foundation.dart';
 import '../../../src/core/models/user.dart';
 
-
 class UserProvider with ChangeNotifier {
-  User _user ;
+  late User _user;
 
   User get getuser => _user;
 
-   setUser(User user) {
+  setUser(User user) {
     print("this is user data1 ${user.full_name}");
     _user = user;
     notifyListeners();
