@@ -6,23 +6,25 @@ import '../../../src/core/config/config.dart';
 
 // The function take will take the user request and verfies it with the api. in this case it will authenticate the user
 Future<dynamic> callAPI(datarequest, myurl, authToken) async {
-  final url = URL + version + myurl;
+  final url = Uri.parse("${URL + version + myurl}");
   print("this is api call $datarequest $url  $authToken");
-  try{final response = await http.post(Uri.parse('$url'),
-      headers: authToken != null
-          ? {
-              HttpHeaders.contentTypeHeader: 'application/json',
-              HttpHeaders.authorizationHeader: "Bearer $authToken"
-            }
-          : {HttpHeaders.contentTypeHeader: 'application/json'},
-      body: json.encode(datarequest));
-  print("this is response of Api call ${json.decode(response.body)}");
-  if (response.statusCode == 200) {
-    print("${response.statusCode}");
-    return json.decode(response.body);
-  } else {
-    throw Exception("Failed to Load Data");
-  }}catch (e) {
+  try {
+    final response = await http.post(Uri.parse('$url'),
+        headers: authToken != null
+            ? {
+                HttpHeaders.contentTypeHeader: 'application/json',
+                HttpHeaders.authorizationHeader: "Bearer $authToken"
+              }
+            : {HttpHeaders.contentTypeHeader: 'application/json'},
+        body: json.encode(datarequest));
+    print("this is response of Api call ${json.decode(response.body)}");
+    if (response.statusCode == 200) {
+      print("${response.statusCode}");
+      return json.decode(response.body);
+    } else {
+      throw Exception("Failed to Load Data");
+    }
+  } catch (e) {
     print("The error$e");
     Map<String, dynamic> response = {
       "statusCode": 400,
@@ -33,24 +35,26 @@ Future<dynamic> callAPI(datarequest, myurl, authToken) async {
 }
 
 Future<dynamic> getAPI(myurl, authToken) async {
-  final url = URL + version + myurl;
+  final url = Uri.parse("${URL + version + myurl}");
   print('this is url $url');
-  try{final response = await http.get(
-    Uri.parse('$url'),
-    headers: authToken != null
-        ? {
-            HttpHeaders.contentTypeHeader: 'application/json',
-            HttpHeaders.authorizationHeader: "Bearer $authToken"
-          }
-        : {HttpHeaders.contentTypeHeader: 'application/json'},
-  );
-  print("this is response of Api call ${json.decode(response.body)}");
-  if (response.statusCode == 200) {
-    print("${response.statusCode}");
-    return json.decode(response.body);
-  } else {
-    throw Exception("Failed to Load Data");
-  }}catch (e) {
+  try {
+    final response = await http.get(
+      Uri.parse('$url'),
+      headers: authToken != null
+          ? {
+              HttpHeaders.contentTypeHeader: 'application/json',
+              HttpHeaders.authorizationHeader: "Bearer $authToken"
+            }
+          : {HttpHeaders.contentTypeHeader: 'application/json'},
+    );
+    print("this is response of Api call ${json.decode(response.body)}");
+    if (response.statusCode == 200) {
+      print("${response.statusCode}");
+      return json.decode(response.body);
+    } else {
+      throw Exception("Failed to Load Data");
+    }
+  } catch (e) {
     print("The error$e");
     Map<String, dynamic> response = {
       "statusCode": 400,
